Issue #SB-25310 fix: Error message gets displayed when the creator creates/updates the course batch details

+++ src/app/client/src/app/modules/learn/components/batch/update-course-batch/update-course-batch.component.ts
@@ -706,17 +706,19 @@ export class UpdateCourseBatchComponent implements OnInit, OnDestroy, AfterViewI
   }
 
   disableDiscussionForum(batchId) {
-    const requestBody = {
-      'sbType': 'batch',
-      'sbIdentifier': batchId,
-      'cid': this.forumIds
-    };
-    this.discussionService.removeForum(requestBody).subscribe(resp => {
-      this.handleInputChange('enable-DF-no');
-      this.toasterService.success(_.get(this.resourceService, 'messages.smsg.m0066'));
-    }, error => {
-      this.toasterService.error(this.resourceService.messages.emsg.m0005);
-    });
+    if (this.forumIds && this.forumIds.length > 0) {
+      const requestBody = {
+        'sbType': 'batch',
+        'sbIdentifier': batchId,
+        'cid': this.forumIds
+      };
+      this.discussionService.removeForum(requestBody).subscribe(resp => {
+        this.handleInputChange('enable-DF-no');
+        this.toasterService.success(_.get(this.resourceService, 'messages.smsg.m0066'));
+      }, error => {
+        this.toasterService.error(this.resourceService.messages.emsg.m0005);
+      });
+    }
   }
 
   handleInputChange(inputId) {
